requirements.txt
--extra-index-url https://gitpro.hanse-merkur.de/api/v4/projects/6644/packages/pypi/simple
pykube-ng==22.1.1
cryptography==36.0.1